Add view() method to TrelliscopeClass
