+++ src/assets/styles/main.css
@@ -72,6 +72,14 @@ body {
   color: #ff5252;
 }
 
+.trelliscope-fullscreen {
+  position: fixed !important;
+  top: 0 !important;
+  left: 0 !important;
+  width: 100vw;
+  height: 100vh;
+}
+
 .react-daterange-picker__clear-button,
 .react-datetimerange-picker__clear-button {
   display: none !important;
