+++ src/components/ContentHeader/ContentHeader.module.scss
@@ -45,7 +45,7 @@
         z-index: 1001;
 
         &Icon {
-          position: fixed;
+          position: absolute;
           cursor: pointer;
           z-index: 1200;
           color: #000;
